Rename store to integrationStore
